⚡️ perf: improve chat workspace performance
