+++ src/store/chat/slices/topic/action.ts
@@ -171,6 +171,8 @@ export const chatTopic: StateCreator<
       [SWR_USE_FETCH_TOPIC, sessionId],
       async ([, sessionId]: [string, string]) => topicService.getTopics({ sessionId }),
       {
+        suspense: true,
+        fallbackData: [],
         onSuccess: (topics) => {
           set({ topics, topicsInit: true }, false, n('useFetchTopics(success)', { sessionId }));
         },
