+++ src/store/tool/slices/store/action.ts
@@ -92,6 +92,7 @@ export const createPluginStoreSlice: StateCreator<
   },
   useFetchInstalledPlugins: () =>
     useSWR<LobeTool[]>(INSTALLED_PLUGINS, pluginService.getInstalledPlugins, {
+      fallbackData: [],
       onSuccess: (data) => {
         set(
           { installedPlugins: data, loadingInstallPlugins: false },
@@ -100,9 +101,12 @@ export const createPluginStoreSlice: StateCreator<
         );
       },
       revalidateOnFocus: false,
+      suspense: true,
     }),
   useFetchPluginStore: () =>
     useSWR<LobeChatPluginsMarketIndex>('loadPluginStore', get().loadPluginStore, {
+      fallbackData: { plugins: [], schemaVersion: 1 },
       revalidateOnFocus: false,
+      suspense: true,
     }),
 });
